plots changes

+++ userfctR/user_model.h
@@ -8,7 +8,7 @@
  *
  * Universite catholique de Louvain, Belgium 
  *
- * Last update : Sun Jan  8 00:09:06 2023
+ * Last update : Sun Jan  8 13:26:31 2023
  * --------------------------------------------------------
  *
  */
@@ -66,7 +66,7 @@ struct UserModel
         int PID;
         int Simple_contact;
         double Belgian_road;
-        int Bumpy;
+        double Bumpy;
     } Status;
  
     struct PID{
